save model weights before evaluation

+++ allennlp/training/checkpointer.py
@@ -90,6 +90,7 @@ class Checkpointer(Registrable):
         epoch: Union[int, str],
         trainer: "allennlp.training.trainer.Trainer",
         is_best_so_far: bool = False,
+        save_model_only=False,
     ) -> None:
         if self._serialization_dir is not None:
             with trainer.get_checkpoint_state() as state:
@@ -97,11 +98,16 @@ class Checkpointer(Registrable):
                 model_path = os.path.join(
                     self._serialization_dir, "model_state_epoch_{}.th".format(epoch)
                 )
-                torch.save(model_state, model_path)
+                if not os.path.isfile(model_path):
+                    torch.save(model_state, model_path)
+                if save_model_only:
+                    return
+
                 training_path = os.path.join(
                     self._serialization_dir, "training_state_epoch_{}.th".format(epoch)
                 )
-                torch.save({**training_states, "epoch": epoch}, training_path)
+                if not os.path.isfile(training_path):
+                    torch.save({**training_states, "epoch": epoch}, training_path)
 
             # The main checkpointing logic is now done, this is just shuffling files around, to keep
             # track of best weights, and to remove old checkpoints, if desired.
